requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
gspread==6.1.2
google-auth==2.35.0
redis==5.0.7
